dao-m8/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.9.5
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
aiosqlite==0.20.0
    # via -r requirements.in
alembic==1.13.1
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
appdirs==1.4.4
    # via yfinance
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via yfinance
certifi==2024.2.2
    # via
    #   curl-cffi
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via curl-cffi
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   duckduckgo-search
    #   nltk
    #   typer
    #   uvicorn
curl-cffi==0.6.3
    # via duckduckgo-search
dataclasses-json==0.6.5
    # via
    #   langchain
    #   langchain-community
dnspython==2.6.1
    # via email-validator
duckduckgo-search==5.3.0
    # via -r requirements.in
email-validator==2.1.1
    # via fastapi
fastapi==0.111.0
    # via
    #   -r requirements.in
    #   fastapi-cli
fastapi-cli==0.0.2
    # via fastapi
frozendict==2.4.2
    # via yfinance
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
html5lib==1.1
    # via yfinance
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
ipfs-api==0.2.3
    # via -r requirements.in
jinja2==3.1.3
    # via fastapi
joblib==1.4.2
    # via nltk
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
jsonschema==4.22.0
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
langchain==0.1.17
    # via -r requirements.in
langchain-community==0.0.36
    # via langchain
langchain-core==0.1.50
    # via
    #   langchain
    #   langchain-community
    #   langchain-text-splitters
langchain-text-splitters==0.0.1
    # via langchain
langsmith==0.1.54
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lxml==5.2.1
    # via yfinance
mako==1.3.3
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
marshmallow==3.21.2
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multitasking==0.0.11
    # via yfinance
mypy-extensions==1.0.0
    # via typing-inspect
nltk==3.8.1
    # via -r requirements.in
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
    #   pandas
    #   yfinance
orjson==3.10.3
    # via
    #   duckduckgo-search
    #   fastapi
    #   langsmith
packaging==23.2
    # via
    #   langchain-core
    #   marshmallow
pandas==2.2.2
    # via yfinance
peewee==3.17.3
    # via yfinance
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via
    #   -r requirements.in
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.18.2
    # via pydantic
pygments==2.17.2
    # via rich
pytelegrambotapi==4.17.0
    # via telebot
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.9
    # via fastapi
pytz==2024.1
    # via
    #   pandas
    #   yfinance
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-core
    #   uvicorn
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.4.28
    # via nltk
requests==2.31.0
    # via
    #   -r requirements.in
    #   ipfs-api
    #   langchain
    #   langchain-community
    #   langsmith
    #   pytelegrambotapi
    #   telebot
    #   yfinance
rich==13.7.1
    # via typer
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   html5lib
    #   ipfs-api
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.29
    # via
    #   -r requirements.in
    #   alembic
    #   langchain
    #   langchain-community
starlette==0.37.2
    # via fastapi
telebot==0.0.5
    # via -r requirements.in
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tqdm==4.66.4
    # via nltk
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.11.0
    # via
    #   aiosqlite
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pandas
ujson==5.9.0
    # via fastapi
urllib3==2.2.1
    # via requests
uvicorn[standard]==0.29.0
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
webencodings==0.5.1
    # via html5lib
websockets==12.0
    # via uvicorn
yarl==1.9.4
    # via aiohttp
yfinance==0.2.38
    # via -r requirements.in
